refactor: move MainAgentConfiguration to test dir, add 3-terminal guide
- Move MainAgentConfiguration from src/main to src/test (flat layout)
- Rewrite README with 3-terminal manual testing guide:
  Terminal 1: Workflow Agent (:8080)
  Terminal 2: Main ReActAgent (:8081, main profile)
  Terminal 3: curl user (trigger → observe interrupt → resume → verify)
- Fix @EnabledIfEnvironmentVariable missing 'matches' attribute
- Fix Message builder compilation (taskId(null) for non-resume calls)
