+++ examples/agent-runtime-a2a-openjiuwen-workflow/src/main/java/com/huawei/ascend/examples/a2a/workflow/QuestionerWorkflowApplication.java
@@ -11,7 +11,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  */
 @SpringBootApplication(scanBasePackages = {
         "com.huawei.ascend.examples.a2a.workflow",
-        "com.huawei.ascend.examples.a2a.workflow.main",
         "com.huawei.ascend.runtime.boot"})
 public class QuestionerWorkflowApplication {
